requirements.txt
flask 
marqo 
pandas  
python-dotenv
tqdm
colorama